feat: add metadata
